+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 2025-00-00 : 0.4.0-beta.6 (prerelease)  
 Update: 
 
+- Added context attributes to the User model
 - Added UserFormResource for populating form data
 - Ported original enraiged-laravel Profiles system into enraiged-core
 - Updated,verified package publishable assets